+++ src/Controller/MusicianController.php
@@ -6,13 +6,14 @@ use App\Service\FileUploader;
 use App\Entity\Musician;
 use App\Form\MusicianType;
 use App\Repository\MusicianRepository;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("/musician")
+ * @Route("/musicians")
  */
 class MusicianController extends AbstractController
 {
@@ -29,13 +30,29 @@ class MusicianController extends AbstractController
     }
 
     /**
+     * @Route("/index", name="admin_musician_index", methods={"GET"})
+     * @param MusicianRepository $musicianRepository
+     * @return Response
+     * @IsGranted("ROLE_ADMIN")
+     */
+    public function adminIndex(MusicianRepository $musicianRepository): Response
+    {
+        return $this->render('admin/musician/index.html.twig', [
+            'musicians' => $musicianRepository->findAll(),
+        ]);
+    }
+
+    /**
      * @Route("/new", name="musician_new", methods={"GET","POST"})
      * @param Request $request
      * @param FileUploader $fileUploader
      * @return Response
+     * @IsGranted("ROLE_ADMIN")
      */
     public function new(Request $request, FileUploader $fileUploader): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $musician = new Musician();
         $form = $this->createForm(MusicianType::class, $musician);
         $form->handleRequest($request);
@@ -57,7 +74,7 @@ class MusicianController extends AbstractController
             return $this->redirectToRoute('musician_index');
         }
 
-        return $this->render('musician/new.html.twig', [
+        return $this->render('admin/musician/new.html.twig', [
             'musician' => $musician,
             'form' => $form->createView(),
         ]);
@@ -71,7 +88,20 @@ class MusicianController extends AbstractController
     public function show(Musician $musician): Response
     {
         return $this->render('musician/show.html.twig', [
-            'musician' => $musician,
+            'musician' => $musician
+        ]);
+    }
+
+    /**
+     * @Route("/show/{id}", name="admin_musician_show", methods={"GET"})
+     * @param Musician $musician
+     * @return Response
+     * @IsGranted("ROLE_ADMIN")
+     */
+    public function adminShow(Musician $musician): Response
+    {
+        return $this->render('admin/musician/show.html.twig', [
+            'musician' => $musician
         ]);
     }
 
@@ -80,9 +110,12 @@ class MusicianController extends AbstractController
      * @param Request $request
      * @param Musician $musician
      * @return Response
+     * @IsGranted("ROLE_ADMIN")
      */
     public function edit(Request $request, Musician $musician): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(MusicianType::class, $musician);
         $form->handleRequest($request);
 
@@ -92,7 +125,7 @@ class MusicianController extends AbstractController
             return $this->redirectToRoute('musician_index');
         }
 
-        return $this->render('musician/edit.html.twig', [
+        return $this->render('admin/musician/edit.html.twig', [
             'musician' => $musician,
             'form' => $form->createView(),
         ]);
@@ -103,9 +136,12 @@ class MusicianController extends AbstractController
      * @param Request $request
      * @param Musician $musician
      * @return Response
+     * @IsGranted("ROLE_ADMIN")
      */
     public function delete(Request $request, Musician $musician): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         if ($this->isCsrfTokenValid('delete'.$musician->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($musician);
